requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
opengradient==0.7.5
sqlalchemy==2.0.30
asyncpg==0.29.0
aiohttp==3.9.5
pytest==8.2.0
pytest-asyncio==0.23.7